chapter21/chapter21-service/src/main/java/org/agoncal/book/javaee7/chapter21/ex15/CardValidator15.java
package org.agoncal.book.javaee7.chapter21.ex15;

import javax.jws.WebService;
import javax.jws.soap.SOAPBinding;
import javax.xml.ws.Endpoint;

import static javax.jws.soap.SOAPBinding.ParameterStyle.WRAPPED;
import static javax.jws.soap.SOAPBinding.Style.RPC;
import static javax.jws.soap.SOAPBinding.Use.ENCODED;
import static javax.jws.soap.SOAPBinding.Use.LITERAL;

/**
 * @author Antonio Goncalves
 *         APress Book - Beginning Java EE 7 with Glassfish 4
 *         http://www.apress.com/
 *         http://www.antoniogoncalves.org
 *         --
 */
@WebService
@SOAPBinding(style = RPC, use = ENCODED)
public class CardValidator15 {

  // ======================================
  // =           Public Methods           =
  // ======================================

  public boolean validate(CreditCard15 creditCard) {

    Character lastDigit = creditCard.getNumber().charAt(creditCard.getNumber().length() - 1);

    if (Integer.parseInt(lastDigit.toString()) % 2 == 0) {
      return true;
    } else {
      return false;
    }
  }

  public static void main(String[] args) {
    Endpoint.publish("http://localhost:8080/cardValidator15", new CardValidator15());
  }
}