chapter21/chapter21-service/src/test/java/org/agoncal/book/javaee7/chapter21/ex15/CardValidator15Test.java
package org.agoncal.book.javaee7.chapter21.ex15;

import org.junit.Test;

import static org.junit.Assert.assertFalse;
import static org.junit.Assert.assertTrue;

/**
 * @author Antonio Goncalves
 *         APress Book - Beginning Java EE 7 with Glassfish 4
 *         http://www.apress.com/
 *         http://www.antoniogoncalves.org
 *         --
 */
public class CardValidator15Test {

  @Test
  public void shouldCheckCreditCardValidity() {

    CreditCard15 creditCard = new CreditCard15("12341234", "12/12", 1234, "VISA");

    CardValidator15 cardValidator = new CardValidator15();
    assertTrue("Credit card should be valid", cardValidator.validate(creditCard));
    creditCard.setNumber("2");
    assertTrue("Credit card should be valid", cardValidator.validate(creditCard));

    creditCard.setNumber("12341233");
    assertFalse("Credit card should not be valid", cardValidator.validate(creditCard));
    creditCard.setNumber("1");
    assertFalse("Credit card should not be valid", cardValidator.validate(creditCard));
  }
}
